+++ hipamd/src/hip_table_interface.cpp
@@ -2011,3 +2011,30 @@ hipError_t hipGraphExecExternalSemaphoresWaitNodeSetParams(
   return hip::GetHipDispatchTable()->hipGraphExecExternalSemaphoresWaitNodeSetParams_fn(
       hGraphExec, hNode, nodeParams);
 }
+hipError_t hipLibraryLoadData(hipLibrary_t* library, const void* code, hipJitOption** jitOptions,
+                              void** jitOptionsValues, unsigned int numJitOptions,
+                              hipLibraryOption** libraryOptions, void** libraryOptionValues,
+                              unsigned int numLibraryOptions) {
+  return hip::GetHipDispatchTable()->hipLibraryLoadData_fn(
+      library, code, jitOptions, jitOptionsValues, numJitOptions, libraryOptions,
+      libraryOptionValues, numLibraryOptions);
+}
+hipError_t hipLibraryLoadFromFile(hipLibrary_t* library, const char* fileName,
+                                  hipJitOption** jitOptions, void** jitOptionsValues,
+                                  unsigned int numJitOptions, hipLibraryOption** libraryOptions,
+                                  void** libraryOptionValues, unsigned int numLibraryOptions) {
+  return hip::GetHipDispatchTable()->hipLibraryLoadFromFile_fn(
+      library, fileName, jitOptions, jitOptionsValues, numJitOptions, libraryOptions,
+      libraryOptionValues, numLibraryOptions);
+}
+hipError_t hipLibraryUnload(hipLibrary_t library) {
+  return hip::GetHipDispatchTable()->hipLibraryUnload_fn(library);
+}
+hipError_t hipLibraryGetKernel(hipKernel_t* pKernel, hipLibrary_t library, const char* name)  {
+  return hip::GetHipDispatchTable()->hipLibraryGetKernel_fn(pKernel, library,
+                                                            name);
+}
+hipError_t hipLibraryGetKernelCount(unsigned int *count, hipLibrary_t library) {
+  return hip::GetHipDispatchTable()->hipLibraryGetKernelCount_fn(count,
+                                                                 library);
+}